rework olink core
- better isolation of remote and client
- better base types for source and sinks
- cleaner structure
