+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = olink-core
+name = olink
 version = 0.0.1
 author = AppiGear
 author_email = info@apigear.io
