+++ src/olink/core/types.py
@@ -1,6 +1,6 @@
 from enum import IntEnum
 from typing import Any, Callable
-from typing import Protocol as ProptocolType
+from typing import Protocol as ProtocolType
 import json
 
 
@@ -74,7 +74,7 @@ class LogLevel:
 WriteLogFunc = Callable[[LogLevel, str], None]
 
 
-class ILogger(ProptocolType):
+class ILogger(ProtocolType):
     def log(level: LogLevel, msg: str) -> None:
         raise NotImplementedError()
