+++ src/olink/ws/__init__.py
@@ -1,0 +1,2 @@
+from .client import Connection
+from .server import Server, run_server
