+++ src/olink/ws/server.py
@@ -1,43 +1,8 @@
 import websockets as ws
-from .emitter import Emitter
 from typing import Any
 import asyncio
-from .client import Client
-from ..remotenode import IObjectSource, RemoteNode
+from olink.remote import RemoteNode
 
-class SourceAdapter(IObjectSource):
-    node: RemoteNode = None
-    object_id: str = None
-    def __init__(self, objectId: str, impl) -> None:
-        self.object_id = objectId
-        self.impl = impl
-        RemoteNode.register_source(self)
-
-    def olink_object_name() -> str:
-        return self.objectId
-    
-    def olink_invoke(self, name: str, args: list[Any]) -> Any:
-        path = Name.path_from_name(name)
-        func = getattr(self.impl, path)
-        try:
-            result = func(**args)
-        except Exception as e:
-            print('error: %s' % e)
-            result = None
-        return result
-    
-    def olink_set_property(self, name: str, value: Any):
-        # set property value on implementation
-        path = Name.path_from_name(name)
-        setattr(self, self.impl, value)
-
-    def olink_linked(self, name: str, node: "RemoteNode"):
-        # called when the source is linked to a client node
-        self.node = node
-
-    def olink_collect_properties(self) -> object:
-        # collect properties from implementation to send back to client node initially
-        return {k: getattr(self.impl, k) for k in ['count']}
 
 
 class RemotePipe:
